+++ api-builder-plugin-fn-redis/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Security
 
+## [0.0.2] 2020-04-13
+- Prevents retry on connect
+- Throw exception in case of Redis connection failure
+- Facade for Redis client
+- Actions use async version of Redis client api
+- Register API Builder hooks for graceful disconnect
+- Reusable utils to register context for actions
+
 ## [0.0.1] 2020-04-08
 ### Added
 - Initial version